demo_flask/requirements.txt
Flask==2.0.2
psycopg2-binary==2.9.3
